(maint) updates... like lots of them
Lots of good stuff

+++ boxes/shared/shell/PreparePuppetProvisioner.ps1
@@ -1,33 +1,37 @@
-﻿$PuppetInstallPath = "$env:SystemDrive\Program Files (x86)\Puppet Labs\Puppet\bin"
-if (!(Test-Path $PuppetInstallPath)) {$PuppetInstallPath = "$env:SystemDrive\Program Files\Puppet Labs\Puppet\bin";}
+﻿$PuppetFile = 'c:\vagrant\puppet\Puppetfile'
 
-$env:Path += ";$PuppetInstallPath"
+if (Test-Path $PuppetFile) {
+    $PuppetInstallPath = "$env:SystemDrive\Program Files (x86)\Puppet Labs\Puppet\bin"
+    if (!(Test-Path $PuppetInstallPath)) {$PuppetInstallPath = "$env:SystemDrive\Program Files\Puppet Labs\Puppet\bin";}
 
-$CommonAppDataPath = [Environment]::GetFolderPath([Environment+SpecialFolder]::CommonApplicationData)
-$PuppetAppDataPath = Join-Path $CommonAppDataPath 'PuppetLabs\puppet'
-$PuppetEtcPath = Join-Path "$PuppetAppDataPath" 'etc'
+    $env:Path += ";$PuppetInstallPath"
 
-if (!(Test-Path $PuppetEtcPath)) {
-  Write-Host "Creating folder `'$PuppetEtcPath`'"
-  $null = New-Item -Path "$PuppetEtcPath" -ItemType Directory
-}
+    $CommonAppDataPath = [Environment]::GetFolderPath([Environment+SpecialFolder]::CommonApplicationData)
+    $PuppetAppDataPath = Join-Path $CommonAppDataPath 'PuppetLabs\puppet'
+    $PuppetEtcPath = Join-Path "$PuppetAppDataPath" 'etc'
 
-Copy-Item "c:\vagrant\puppet\Puppetfile" "$PuppetEtcPath"
+    if (!(Test-Path $PuppetEtcPath)) {
+      Write-Host "Creating folder `'$PuppetEtcPath`'"
+      $null = New-Item -Path "$PuppetEtcPath" -ItemType Directory
+    }
 
-Push-Location "$PuppetEtcPath"
-Write-Host "Ensuring librarian puppet"
-try {
-    $ErrorActionPreference = "Stop";
-    Get-Command librarian-puppet
-    Write-Host "Running librarian-puppet update"
-    & librarian-puppet update
-} catch {
-    Write-Host "Installing librarian puppet"
-    & gem install librarian-puppet --no-ri --no-rdoc
-    Write-Host "Running librarian-puppet install --clean"
+    Copy-Item "$PuppetFile" "$PuppetEtcPath"
 
-    & librarian-puppet install --clean
-} finally {
-    $ErrorActionPreference = "Continue";
-    Pop-Location
+    Push-Location "$PuppetEtcPath"
+    Write-Host "Ensuring librarian puppet"
+    try {
+        $ErrorActionPreference = "Stop";
+        Get-Command librarian-puppet
+        Write-Host "Running librarian-puppet update"
+        & librarian-puppet update
+    } catch {
+        Write-Host "Installing librarian puppet"
+        & gem install librarian-puppet --no-ri --no-rdoc
+        Write-Host "Running librarian-puppet install --clean"
+
+        & librarian-puppet install --clean
+    } finally {
+        $ErrorActionPreference = "Continue";
+        Pop-Location
+    }
 }
